feat(manage): make update a case more generic
feat(manage): make update a case more generic

fix: wip
